requirements.txt
dbt-core==1.8.2
python-dotenv==1.0.1
dbt-sqlserver==1.8.7